# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Category:
    properties:
      business_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.CreateCategoryRequestForSwagger:
    properties:
      name:
        type: string
    type: object
...
  entity.GetAllCategoriesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.Category'
        type: array
      total:
        type: integer
    type: object
...
  entity.UpdateCategoryRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  status_http.Response:
    properties:
      custom_message: {}
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /category/create:
    post:
      consumes:
      - application/json
      description: Create a new category in the database
      parameters:
      - description: Category Details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.CreateCategoryRequestForSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - CATEGORY
  /category/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category from the database by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request Data
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Category Not Found
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a category by ID
      tags:
      - CATEGORY
  /category/get/{id}:
    get:
      consumes:
      - application/json
      description: Get category details by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Category'
              type: object
        "400":
          description: Bad request data
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Category Not Found
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get category by ID
      tags:
      - CATEGORY
  /category/list:
    get:
      consumes:
      - application/json
      description: Get a list of categories with optional filters and pagination
      parameters:
      - description: Filter by Category Name
        in: query
        name: name
        type: string
      - description: 'Limit the number of results (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of Categories
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.GetAllCategoriesResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - CATEGORY
  /category/update:
    put:
      consumes:
      - application/json
      description: Update category details by ID
      parameters:
      - description: Category Details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad request data
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Category Not Found
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing category
      tags:
      - CATEGORY
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"